Feature: Refactored CSV reader into Library and Made two methods to read All and 1 Record to end
