+++ csv.sln
@@ -13,6 +13,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "csvfiles", "csvfiles", "{B0
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "csvTests", "csvTests\csvTests.csproj", "{7D4ECEBF-1B1F-4B5B-B455-9E236476665E}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Worldcities", "Worldcities\Worldcities.csproj", "{ADDBA089-5FA6-4EB5-84D1-43CB7C1CE39B}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -27,6 +29,10 @@ Global
 		{7D4ECEBF-1B1F-4B5B-B455-9E236476665E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7D4ECEBF-1B1F-4B5B-B455-9E236476665E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7D4ECEBF-1B1F-4B5B-B455-9E236476665E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{ADDBA089-5FA6-4EB5-84D1-43CB7C1CE39B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{ADDBA089-5FA6-4EB5-84D1-43CB7C1CE39B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{ADDBA089-5FA6-4EB5-84D1-43CB7C1CE39B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{ADDBA089-5FA6-4EB5-84D1-43CB7C1CE39B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
