+++ csv/EmptyDouble.cs
@@ -1,0 +1,25 @@
+﻿using CsvHelper;
+using CsvHelper.Configuration;
+using CsvHelper.TypeConversion;
+using System.Collections.Generic;
+using System.IO;
+using Worldcities;
+namespace csv
+{
+    public class EmptyDouble : DoubleConverter
+    {
+        public override object ConvertFromString(string text, IReaderRow row, MemberMapData memberMapData)
+        {
+            if (text == "")
+            {
+                double value = 0;
+                return value;
+            }
+            else
+            {
+                double value = System.Convert.ToDouble(text);
+                return value;
+            }
+        }
+    }
+}
