+++ csv/csv.csproj
@@ -8,4 +8,8 @@
     <PackageReference Include="CsvHelper" Version="12.3.2" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\Worldcities\Worldcities.csproj" />
+  </ItemGroup>
+
 </Project>
