+++ csvTests/readcsvTests.cs
@@ -9,11 +9,22 @@ namespace csv.Tests
     [TestClass()]
     public class readcsvTests
     {
+        
+
         [TestMethod()]
-        public void MainTest()
+        public void ReadInCSVTest()
         {
-            readcsv.Main();
-            Assert.Fail();
+            string path = "c://csvfiles//worldcities.csv";
+            List<Worldcities.City> cities = readcsv.ReadAllInCSV(path);
+            Assert.AreEqual(15493, cities.Count);
+        }
+
+        [TestMethod()]
+        public void ReadInCSV1Test()
+        {
+            string path = "c://csvfiles//worldcities.csv";
+            List<Worldcities.City> cities = readcsv.ReadAllInCSV(path);
+            Assert.AreEqual(15493, cities.Count);
         }
     }
 }
